requirements.txt
RPI.GPIO
adafruit-blinka
circuitpython-build-tools
adafruit-circuitpython-lis3dh
adafruit-circuitpython-pca9685
numpy
pillow
obd
picamera
alembic==1.0.11
asn1crypto==0.24.0
cffi==1.12.3
clg==2.0.0
Click==7.0
cryptography==2.7
dnspython==1.16.0
docopt==0.6.2
dominate==2.3.5
email-validator==1.0.4
enum34==1.1.6
Flask==1.0.3
Flask-Admin==1.5.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-MySQL==1.4.0
flask-nav==0.6
Flask-SQLAlchemy==2.4.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
loadconfig==0.1.1
Mako==1.0.13
MarkupSafe==1.1.1
migrant==1.3.1
mysql-connector==2.2.9
peewee==3.9.6
peewee-migrations==0.3.18
pycparser==2.19
PyMySQL==0.9.3
python-dateutil==2.8.0
python-editor==1.0.4
PyYAML==5.1.1
six==1.12.0
SQLAlchemy==1.3.5
termcolor==1.1.0
visitor==0.1.3
Werkzeug==0.15.4
wtf-peewee==3.0.0
WTForms==2.2.1
XlsxWriter==1.1.8
